WIP: swap out prosemirror and get text highlighting working. doesn't parse and persist to supabase yet. will be a merge of all thoughts into one text editor
