+++ package.json
@@ -10,12 +10,21 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@remirror/pm": "^2.0.8",
+    "@remirror/react": "^2.0.35",
+    "@remirror/react-editors": "^1.0.40",
     "@supabase/supabase-js": "^2.39.3",
     "draft-js": "^0.11.7",
     "openai": "^4.27.0",
+    "prosemirror-model": "^1.19.4",
+    "prosemirror-schema-basic": "^1.2.2",
+    "prosemirror-state": "^1.4.3",
+    "prosemirror-view": "^1.32.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scroll-to-bottom": "^4.2.0"
+    "react-scroll-to-bottom": "^4.2.0",
+    "remirror": "^2.0.38",
+    "use-prosemirror": "^1.3.0"
   },
   "devDependencies": {
     "@types/draft-js": "^0.11.17",
